Revert "Move Tika substitutions to proper package"
This reverts commit 462eff2daa3c09fa6cf5bd514b34a486fcf947fb.
